% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allWQ}
\alias{allWQ}
\title{Lake Michigan water quality data compiled across federal surveys.}
\format{
A data frame with 374368 rows and 27 variables:
\describe{
\item{UID}{Unique sample ID across all sources}
\item{Study}{Name of study}
\item{SITE_ID}{Name of site}
\item{Latitude}{Latitude in decimal degrees}
\item{Longitude}{Longitude in decimal degrees}
\item{stationDepth}{Station depth in meters}
\item{sampleDate}{Date of sampling event. This is either the date provided with the data (with unknown time of day) or is the date associated with the collection time in EST (and thus may differ from the date in sampleDateTimeUTC)}
\item{sampleDateTimeUTC}{Date-time of sampling event in UTC time zone, where available}
\item{sampleDepth}{Depth of sample in meters}
\item{DEPTH_CODE}{Depth code of sample, where available}
\item{CodeName}{Harmonized short name for the water quality parameter}
\item{LongName}{Harmonized descriptive name for the water quality parameter}
\item{Category}{Water quality category inspired by LAGOS-US LIMNO. https://doi.org/10.6073/pasta/2c58f5a50ab813919f99cc1f265f271c}
\item{ANALYTE_Orig_Name}{Analyte name in the original data source}
\item{RESULT}{Value of the measurement}
\item{MDL}{Method detection limit, where available}
\item{RL}{Reporting limit, where available}
\item{Units}{Units of measurement for RESULT, MDL and RL}
\item{ReportedUnits}{Measurement units in the original data source}
\item{ConversionFactor}{Multiplicative conversion factor if conversion was necessary to convert from ReportedUnits to Units}
\item{Unified_Flag}{Unified quality control flag across datasets}
\item{Unified_Comment}{Unified quality control comment across datasets}
\item{METHOD}{Method of measurement, where available}
\item{LAB}{Lab taking measurement, where available}
\item{Orig_QAcode}{QA code in the original data source or added in data checks}
\item{Orig_QAcomment}{QA comment in the original data source or added in data checks}
\item{Orig_QAdefinition}{QA definition in the original data source}
}
}
\source{
GLharmonizeR Great Lakes water quality R package. See GitHub documentation for more information (https://github.com/USEPA/GLharmonizeR).
}
\usage{
allWQ
}
\description{
A dataset containing water quality measurements from NCCA, CSMI, GLNPO,
and NOAA data sources. See GitHub documentation for more information (https://github.com/USEPA/GLharmonizeR).
}
\author{
Kelsey Vitense \email{vitense.kelsey@epa.gov}

Christian Coffman
}
\keyword{datasets}
