% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFullData.R
\name{assembleData}
\alias{assembleData}
\title{Load and join all WQ data from 2010, 2015, and 2020/2021 from CSMI, GLENDA, and NCCA}
\usage{
assembleData(out = NULL, .test = FALSE, binaryOut = FALSE)
}
\arguments{
\item{out}{(optional) filepath to save the dataset to}

\item{.test}{(optional) boolean, if testing that data loads and joins, this flag only loads
parts of the datasets to test it faster}

\item{binaryOut}{(optional) boolean, should saved data be RDS format for efficiency?}
}
\value{
dataframe of the fully joined water quality data from CSMI, NCCA, and GLENDA over years 2010, 2015, 2021
}
\description{
\code{.LoadJoinAll} returns a dataframe data from 2010, 2015, and 2020/2021 from CSMI, GLENDA, and NCCA'
}
\details{
This is a hidden function that should not generally be used by users.
}
