% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLENDA.R
\name{.cleanGLENDA}
\alias{.cleanGLENDA}
\title{cleanGLENDA}
\usage{
.cleanGLENDA(
  df,
  namingFile,
  imputeCoordinates = TRUE,
  GLENDAsitePath = NULL,
  GLENDAlimitsPath
)
}
\arguments{
\item{df}{GLENDA dataframe in long format, output from .readFormatGLENDA}

\item{namingFile}{filepath to a file containing mappings for analyte names}

\item{imputeCoordinates}{(optional) Boolean specifying whether to impute missing station coordinates,}

\item{GLENDAsitePath}{filepath to list of site coordinates to fill in missing lats/lons}

\item{GLENDAlimitsPath}{filepath to a file with detection limits for GLENDA data}
}
\value{
a dataframe
}
\description{
A function to perform QC on the GLENDA data
}
\details{
\code{.cleanGLENDA} This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
