% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmi2010.R
\name{.LoadCSMI2010}
\alias{.LoadCSMI2010}
\title{Load and join data for CSMI 2010 from csv excel files}
\usage{
.LoadCSMI2010(directoryPath)
}
\arguments{
\item{directorypath}{a string specifying the directory path of the access database}
}
\value{
dataframe of the fully joined water quality data from CSMI 2010
}
\description{
\code{.LoadCSMI2010} returns a dataframe of all of the joined water quality data relating to CSMI 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
