% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmi2015.R
\name{.LoadCSMI2015}
\alias{.LoadCSMI2015}
\title{Load and join data for CSMI 2015 from access database}
\usage{
.LoadCSMI2015(filepath)
}
\arguments{
\item{filepath}{a string specifying the filepath of the access database}
}
\value{
dataframe of the fully joined water quality data from CSMI 2015
}
\description{
\code{.LoadCSMI2015} returns a dataframe of all of the joined water quality data relating to CSMI 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
