% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAfulljoin.R
\name{.loadNCCA}
\alias{.loadNCCA}
\title{Read in NCCA 2010 and 2015 water chemistry and hydrographic data}
\usage{
.loadNCCA(
  NCCAsites2010,
  NCCAsites2015,
  NCCAwq2010,
  NCCAwq2015,
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015,
  namingFile,
  Lakes = NULL,
  n_max = Inf
)
}
\arguments{
\item{NCCAsites2010}{a string specifying the URL for the 2010 site data}

\item{NCCAsites2015}{a string specifying the URL for the 2015 site data}

\item{NCCAwq2010}{a string specifying the URL for the 2010 water chemistry data}

\item{NCCAwq2015}{a string specifying the URL for the 2015 water chemistry data}

\item{NCCAhydrofiles2010}{a string specifying the URL for the 2010 hydrographic data}

\item{NCCAhydrofile2015}{a string specifying the URL for the 2015 hydrographic data}

\item{NCCAsecchifile2015}{a string specifying the URL for the 2015 Secchi data}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{Lakes}{List of Great Lakes to output}

\item{n_max}{integer specifying how many lines to read of each file to save time for testing}
}
\value{
dataframe
}
\description{
\code{.loadNCCA} returns water quality data along with spatial data from the
site information measured through NCCA study in 2010 and 2015
}
\details{
Join the water quality and hydrographic data and filter to specified lakes.
}
