% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAhydro2010}
\alias{.loadNCCAhydro2010}
\title{Load and join NCCA 2010 hydrographic data from online csv files}
\usage{
.loadNCCAhydro2010(
  NCCAhydrofiles2010,
  NCCAsites2010,
  namingFile,
  n_max = n_max
)
}
\arguments{
\item{NCCAhydrofiles2010}{a string specifying the URL for the data}

\item{NCCAsites2010}{a string specifying the URL for the site data}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{n_max}{Number of rows to read in from the data file (this is just for testing purposes)}
}
\value{
dataframe
}
\description{
\code{.loadNCCAhydro2010} returns a dataframe of all of the hydrographic data relating to NCCA 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
