% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAhydro2020}
\alias{.loadNCCAhydro2020}
\title{Load and join NCCA 2020 hydrographic data from csv files}
\usage{
.loadNCCAhydro2020(NCCAhydrofile2020, NCCAsites2020, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAhydrofile2020}{a string specifying the filepath of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCAhydro2020} returns a dataframe of all of the hydrographic data relating to NCCA 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
