% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAsites.R
\name{.loadNCCASite2010}
\alias{.loadNCCASite2010}
\title{Read in NCCA site data}
\usage{
.loadNCCASite2010(NCCAsites2010)
}
\arguments{
\item{NCCAsites2010}{a string specifying the filepath of the data. This is usually a url.}
}
\value{
dataframe
}
\description{
\code{.loadNCCASite2010} returns spatial data relating to study sites
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
