% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAsites.R
\name{.loadNCCASite2015}
\alias{.loadNCCASite2015}
\title{Read in NCCA site data}
\usage{
.loadNCCASite2015(NCCAsites2015)
}
\arguments{
\item{NCCAsites2015}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCASite2015} returns spatial data relating to study sites
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
