% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.loadNCCAwq2010}
\alias{.loadNCCAwq2010}
\title{Read in all NCCA water chemistry from 2010}
\usage{
.loadNCCAwq2010(NCCAwq2010, NCCAsites2010, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAwq2010}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCAwq2010} returns water chemistry data measured during the NCCA study in 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
