% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.loadNCCAwq2020}
\alias{.loadNCCAwq2020}
\title{Read in all NCCA water chemistry from 2020}
\usage{
.loadNCCAwq2020(NCCAwq2020, NCCAsites2020, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAwq2020}{a string specifying the directory of the data}

\item{NCCAsites2020}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCAwq2020} returns water chemistry data measured during the NCCA study in 2020
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
