% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAAwq.R
\name{.loadNOAAwq}
\alias{.loadNOAAwq}
\title{Read in NOAA water chemistry data}
\usage{
.loadNOAAwq(noaaWQ, noaaWQ2, namingFile, noaaWQSites)
}
\arguments{
\item{noaaWQ}{a string specifying the URL for the corrected data based on 'edits_nutrients.xlsx' sent 2024-12-13. These were manually corrected.}

\item{noaaWQ2}{a string specifying the URL for additional nutrient data provided by GLERL.}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{noaaWQSites}{a string specifying the URL for the site data}
}
\value{
dataframe
}
\description{
\code{.loadNOAAwq} returns water chemistry data provided by NOAA GLERL
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
