% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterQuality.R
\name{.parseNCCAData}
\alias{.parseNCCAData}
\title{Parse NCCA data files stored in a common directory}
\usage{
.parseNCCAData(directory)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.parseNCCAData} returns filepaths of data NCCA stored files
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
