% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmi2021.R
\name{.readCleanCSMI2021}
\alias{.readCleanCSMI2021}
\title{Load and join data for CSMI 2021 from csv excel files}
\usage{
.readCleanCSMI2021(csmi2021, namingFile)
}
\arguments{
\item{csmi2021}{a string specifying the directory to CSMI 2021 data}
}
\value{
dataframe of the fully joined water quality data from CSMI 2021
}
\description{
\code{.readCleanCSMI2021} returns a dataframe of all of the joined water quality data relating to CSMI 2021
}
\details{
this function implicitly when assembling their full water quality dataset
}
