% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLENDA.R
\name{.readFormatGLENDA}
\alias{.readFormatGLENDA}
\title{readPivotGLENDA}
\usage{
.readFormatGLENDA(Glenda, n_max = Inf, sampleIDs = NULL)
}
\arguments{
\item{Glenda}{a filepath to the GLENDA RDS or CSV file}

\item{n_max}{Number of rows to read in from the data file (this is just for testing purposes)}

\item{sampleIDs}{a list of sampleIDs to keep in the final dataset}
}
\value{
a dataframe
}
\description{
A function to read the full GLENDA data file downloaded 2023/10/23 and convert to long format.
}
\details{
\code{.readFormatGLENDA} This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset.
}
