% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterQuality.R
\name{.readNCCA2000s}
\alias{.readNCCA2000s}
\title{Read in all NCCA water quality from the early 2000s}
\usage{
.readNCCA2000s(filepath)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCA2000s} returns water quality data measured during the NCCA study in the early 2000s
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
