% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.readNCCA2010}
\alias{.readNCCA2010}
\title{Read in all NCCA water quality from 2010}
\usage{
.readNCCA2010(filepaths, n_max = Inf)
}
\arguments{
\item{filepaths}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCA2010} returns water quality data measured during the NCCA study in 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
