% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterQuality.R
\name{.readNCCA2015}
\alias{.readNCCA2015}
\title{Read in all NCCA water quality from 2015}
\usage{
.readNCCA2015(filepath)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCA2015} returns water quality data measured during the NCCA study in 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
