% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.readNCCAchemistry}
\alias{.readNCCAchemistry}
\title{Read in all NCCA from 2000s, 2010, and 2015}
\usage{
.readNCCAchemistry(
  tenFiles = NULL,
  fifteenFiles = NULL,
  nccaWQqaFile = NULL,
  n_max = n_max
)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCAchemistry} returns water quality data along with spatial data from the
site information measured through NCCA study in the early 2000s as well as in 2010, and 2015
}
\details{
The spatial information for sites is read in using the .readSites helper functions, this is then
joined to the water quality and ultimately output as a data table.
}
