% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.readNCCAhydro}
\alias{.readNCCAhydro}
\title{Load and join hydrographic and secchi data for NCCA 2010 and 2015}
\usage{
.readNCCAhydro(
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015,
  NCCAwqQA,
  n_max = Inf
)
}
\arguments{
\item{filepath}{a string specifying the filepath of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCAhydro} returns a dataframe of all of the hydrographic data relating to NCCA 2010 and 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
