% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.readNCCAhydro2010}
\alias{.readNCCAhydro2010}
\title{Load and join secchi data for NCCA 2010 hydrographic data from csv files}
\usage{
.readNCCAhydro2010(NCCAhydrofiles2010, n_max = n_max)
}
\arguments{
\item{filepath}{a string specifying the filepath of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCAhydro2010} returns a dataframe of all of the hydrographic data relating to NCCA 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
