% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAsites.R
\name{.readNCCASite2020}
\alias{.readNCCASite2020}
\title{Read in NCCA site data}
\usage{
.readNCCASite2020(NCCAsites2020)
}
\arguments{
\item{NCCAsites2020}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCASite2020} returns spatial data relating to study sites
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
