% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLENDA.R
\name{.readPivotGLENDA}
\alias{.readPivotGLENDA}
\title{readPivotGLENDA}
\usage{
.readPivotGLENDA(filepath)
}
\arguments{
\item{filepath}{a filepath to the GLENDA csv}
}
\value{
a dataframe
}
\description{
A function to read the full GLENDA csv file and convert it to a more user friendly long format.
}
\details{
\code{.readPivotGLENDA} This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
