% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterQuality.R
\name{.readSite}
\alias{.readSite}
\title{Read in NCCA site data}
\usage{
.readSite(filepath)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readSite} returns spatial data relating to study sites
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
