% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFullData.R
\name{.UnifyUnitsNames}
\alias{.UnifyUnitsNames}
\title{Unify the names and units of the fully joined data}
\usage{
.UnifyUnitsNames(data, namingFile)
}
\arguments{
\item{namingFile}{a string specifying the file containing naming and units information}
}
\value{
dataframe with unified names and units
}
\description{
\code{.UnifyUnitsNames} returns a dataframe data from 2010, 2015, and 2020/2021 from CSMI, GLENDA, and NCCA'
}
\details{
This is a hidden function that should not generally be used by users.
}
