% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLharmonizeR-package.R
\docType{package}
\name{GLharmonizeR-package}
\alias{GLharmonizeR}
\alias{GLharmonizeR-package}
\title{GLharmonizeR: Great Lakes water quality data harmonizer}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package contains functions to integrate Great Lakes water quality data from different federal sources, with data currently available for Lake Michigan. The R package harmonizes water quality parameter names, units, and quality control (QC) flags across survey efforts. Functions for imputing missing and censored data and linking observations in time and space are in progress. This R package and associated dataset are intended to improve the ability of investigators in federal, state, and local agencies and academia to utilize Great Lakes water quality data to better understand and manage the Great Lakes.
}
\author{
\strong{Maintainer}: Kelsey Vitense \email{vitense.kelsey@epa.gov} (\href{https://orcid.org/0000-0003-2455-5513}{ORCID})

Authors:
\itemize{
  \item Christian Coffman (\href{https://orcid.org/0000-0002-5524-864X}{ORCID})
}

}
\keyword{internal}
