% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMChla-package.R
\docType{package}
\name{LMChla-package}
\alias{LMChla}
\alias{LMChla-package}
\title{LMChla: Lake Michigan Chlorophyll-a and water quality assembler}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package loads in-situ water quality data from NCCA, CSMI, and GLENDA covering the years 2010, 2015, and 2020/2021, plus some additional data in a number of other years comprised of both CTD and lab sample data. It standardizes the data for joining and runs quality control checks on each of the data sources separately as well as in its integrated state. It is meant to lower the barrier of entry for researchers interested in water quality in the great lakes over this time period.
}
\author{
\strong{Maintainer}: Kelsey Vitense \email{vitense.kelsey@epa.gov}

Authors:
\itemize{
  \item Christian Coffman \email{coffman.christian@epa.gov} (\href{https://orcid.org/0000-0002-5524-864X}{ORCID})
}

}
\keyword{internal}
