% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAfulljoin.R
\name{LoadNCCAfull}
\alias{LoadNCCAfull}
\title{Read in all NCCA from 2000s, 2010, and 2015 including hydrogrpahic data}
\usage{
LoadNCCAfull(
  siteFiles,
  preFiles = NULL,
  tenFiles = NULL,
  fifteenFiles = NULL,
  greatLakes = TRUE,
  Lakes = NULL,
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015
)
}
\arguments{
\item{siteFiles}{filepath to site files}

\item{preFiles}{filepath to pre 2010's data}

\item{tenFiles}{filepath to 2010's data}

\item{fifteenFiles}{filepaht to 2015 data}

\item{NCCAhydrofiles2010}{filepath to hydrogrpahic 2010 data}

\item{NCCAhydrofile2015}{filepath to hydrogarphic 2015 data}

\item{NCCAjsecchifile2015}{filepaht to secchi 2015 data}
}
\value{
dataframe
}
\description{
\code{LoadNCCAfull} returns water quality data along with spatial data from the
site information measured through NCCA study in the early 2000s as well as in 2010, and 2015
}
\details{
The spatial information for sites is read in using the .readSites helper functions, this is then
joined to the water quality and hydrographic data and ultimately output as a data table.
}
