% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFullData.R
\name{LoadWQdata}
\alias{LoadWQdata}
\title{Read and unite water quality data from CSMI, NCCA, and GLENDA across 2010, 2015, and 2020/2021}
\usage{
LoadWQdata(
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015,
  siteFiles,
  preFiles,
  tenFiles,
  fifteenFiles,
  glendaData,
  csmi2010,
  csmi2015,
  csmi2021,
  namingFile
)
}
\arguments{
\item{siteFiles}{a string specifying the directory containing NCCA Site data}

\item{preFiles}{a string specifying the path containing 2010's and earlier WQ data}

\item{tenFiles}{a string specifying the directory containing NCCA WQ 2010 data}

\item{fifteenFiles}{a string specifying the directory containing NCCA wQ 2015 data}

\item{glendaData}{a string specifying the directory containing GLENDA data}

\item{csmi2010}{a string specifying the directory containing CSMI 2010 data}

\item{csmi2015}{a string specifying the directory containing CSMI 2015 data}

\item{csmi2021}{a string specifying the directory containing CSMI 2021 data}

\item{namingFile}{a string specifying the file containing naming and units information}
}
\value{
dataframe with unified names and units for all WQ data
}
\description{
\code{LoadWQdata} is the main function of this package and returns a dataframe data from 2010, 2015,
and 2020/2021 from CSMI, GLENDA, and NCCA'
}
\details{
Loading data from all of these sources required QC on each respective source.
}
