% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLENDA.R
\name{readCleanGLENDA}
\alias{readCleanGLENDA}
\title{readCleanGLENDA}
\usage{
readCleanGLENDA(
  filepath,
  flagsPath = NULL,
  siteCoords = NULL,
  imputeCoordinates = FALSE,
<<<<<<< HEAD
  nameMap = NULL
=======
  nameMap = NULL,
  n_max = Inf,
  sampleIDs = NULL
>>>>>>> 38-tests-for-data-quality
)
}
\arguments{
\item{filepath}{a filepath to the GLENDA csv}
}
\value{
a dataframe
}
\description{
A function to read and clean the full GLENDA csv file. This is generally the
function users will interact with, the comprising read and clean functions
are moreso for development purposes.
}
