% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterQuality.R
\name{readNCCA}
\alias{readNCCA}
\title{Read in all NCCA from 2000s, 2010, and 2015}
\usage{
readNCCA(siteFiles, preFiles = NULL, tenFiles = NULL, fifteenFiles = NULL)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{readNCCA} returns water quality data along with spatial data from the
site information measured through NCCA study in the early 2000s as well as in 2010, and 2015
}
\details{
The spatial information for sites is read in using the .readSites helper functions, this is then
joined to the water quality and hydrographic data and ultimately output as a data table.
}
