% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFullData.R
\name{assembleData}
\alias{assembleData}
\title{Load and join Great Lakes water quality data from GLNPO, CSMI, NCCA, and NOAA. Data currently limited to Lake Michigan.}
\usage{
assembleData(out = NULL, .test = FALSE, binaryOut = TRUE)
}
\arguments{
\item{out}{filepath specifying where to save the assembled data. Note: this should exclude the file extension}

\item{.test}{(optional) boolean, load a test subset of the data. Speeds up function for developers}

\item{binaryOut}{(optional) boolean, should saved data be RDS format for efficiency?. If false it is saved as CSV Default is TRUE.}
}
\value{
Harmonized water quality dataset for Lake Michigan. See GitHub page for details on data sources, water quality parameter names, and quality control flags.
}
\description{
\code{assembleData} returns a dataframe with harmonized water quality parameter names, units, and quality control flags
}
\details{
This is the main function of GLharmonizeR which assembles water quality data for the Great Lakes (currently only Lake Michigan) across multiple decades and federal data sources.
}
\examples{
df <- assembleData() # Creates object df for use in current R session
df <- assembleData("filepath", binaryOut = FALSE) # Same as above but also write data to "filepath.rda"
df <- assembleData("filepath", binaryOut = TRUE) # Same as above but  write data to "filepath.csv"
}
