% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinFullData.R
\name{assembleData}
\alias{assembleData}
\title{Load and join all WQ data from 2010, 2015, and 2020/2021 from CSMI, GLENDA, and NCCA}
\usage{
assembleData(
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015,
  ncca2010sites,
  ncca2015sites,
  tenFiles,
  tenQAfile,
  fifteenFiles,
  glendaData,
  csmi2010,
  csmi2015,
  csmi2021,
  seaBirdFiles,
  namingFile,
  out = NULL,
  test = FALSE,
  binaryOut = FALSE
)
}
\arguments{
\item{tenFiles}{a string specifying the directory containing NCCA WQ 2010 data}

\item{fifteenFiles}{a string specifying the directory containing NCCA wQ 2015 data}

\item{glendaData}{a string specifying the directory containing GLENDA data}

\item{csmi2010}{a string specifying the directory containing CSMI 2010 data}

\item{csmi2015}{a string specifying the directory containing CSMI 2015 data}

\item{csmi2021}{a string specifying the directory containing CSMI 2021 data}

\item{out}{(optional) filepath to save the dataset to}

\item{test}{(optional) boolean, if testing that data loads and joins, this flag only loads
parts of the datasets to test it faster}

\item{hydrofiles2010}{a string specifying the directory containing NCCA hydrographic 2010 data}

\item{hydrofile2015}{a string specifying the directory containing NCCA hydrographic 2015 data}

\item{secchifile2015}{a string specifying the directory containing NCCA secchi 2015 data}

\item{siteFiles}{a string specifying the directory containing NCCA Site data}

\item{preFiles}{a string specifying the path containing 2010's and earlier WQ data}
}
\value{
dataframe of the fully joined water quality data from CSMI, NCCA, and GLENDA over years 2010, 2015, 2021
}
\description{
\code{.LoadJoinAll} returns a dataframe data from 2010, 2015, and 2020/2021 from CSMI, GLENDA, and NCCA'
}
\details{
This is a hidden function that should not generally be used by users.
}
