% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLENDA.R
\name{.cleanGLENDA}
\alias{.cleanGLENDA}
\title{cleanGLENDA}
\usage{
.cleanGLENDA(
  df,
  namingFile,
  flagsPath = NULL,
  imputeCoordinates = FALSE,
  siteCoords = NULL,
  limitsPath = NULL
)
}
\arguments{
\item{df}{GLENDA dataframe in long format}

\item{namingFile}{(optional) filepath to a file containing remappings for analyte names}

\item{flagsPath}{(optional) filepath to the Result remarks descriptions. Default is NULL.}

\item{siteCoords}{(optional) filepath to list of site coordinates to impute missing lats/lons}

\item{imputeCoordinages}{(optional) Boolean specifying whether to impute missing station coordinates}
}
\value{
a dataframe
}
\description{
A function to perform automated QC on the GLENDA data
}
\details{
\code{.cleanGLENDA} This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
