% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmiJoinAll.R
\name{.LoadCSMI}
\alias{.LoadCSMI}
\title{Load and join CSMI water quality data from 2010, 2015, and 2021.}
\usage{
.LoadCSMI(csmi2010, csmi2021, namingFile, n_max = Inf)
}
\arguments{
\item{csmi2010}{a string specifying the directory containing CSMI 2010 data}

\item{csmi2021}{a string specifying the directory containing the CSMI 2021 data}
}
\value{
dataframe of the fully joined water quality data from CSMI years 2010, 2015, 2021
}
\description{
\code{LoadCSMI} returns a dataframe of all of the joined water quality data relating to CSMI years
2010, 2015, and 2021.
}
\details{
This is the main functions users should use to load and assemble CSMI data
using this package. This function is also called in over arching functions
to assemble data across multiple data sources.
}
