% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmi2021.R
\name{.LoadCSMI2021}
\alias{.LoadCSMI2021}
\title{Load and join data for CSMI 2021 from csv excel files}
\usage{
.LoadCSMI2021(csmi2021, n_max = Inf)
}
\arguments{
\item{csmi2021}{a string specifying the directory to CSMI 2021 data}
}
\value{
dataframe of the fully joined water quality data from CSMI 2021
}
\description{
\code{.LoadCSMI2021} returns a dataframe of all of the joined water quality data relating to CSMI 2021
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
