% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAfulljoin.R
\name{.loadNCCA}
\alias{.loadNCCA}
\title{Read in all NCCA from 2000s, 2010, and 2015 including hydrogrpahic data}
\usage{
.loadNCCA(
  NCCAsites2010,
  NCCAsites2015,
  NCCAwq2010,
  NCCAwq2015,
  NCCAhydrofiles2010,
  NCCAhydrofile2015,
  NCCAsecchifile2015,
  namingFile,
  Lakes = NULL,
  n_max = Inf
)
}
\arguments{
\item{NCCAsites2010}{filepath to site files}

\item{NCCAsites2015}{filepath to site files}

\item{NCCAwq2010}{filepath to 2010's data}

\item{NCCAwq2015}{filepaht to 2015 data}

\item{NCCAhydrofiles2010}{filepath to hydrogrpahic 2010 data}

\item{NCCAhydrofile2015}{filepath to hydrogarphic 2015 data}

\item{namingFile}{filepath to Analytes3.xlsx which conatains names and conversions}

\item{Lakes}{List of Lakes to output}

\item{n_max}{integer specifying how many lines to read of each file to save time for testing}

\item{NCCAjsecchifile2015}{filepaht to secchi 2015 data}
}
\value{
dataframe
}
\description{
\code{.loadNCCA} returns water quality data along with spatial data from the
site information measured through NCCA study in the early 2000s as well as in 2010, and 2015
}
\details{
The spatial information for sites is read in using the .readSites helper functions, this is then
joined to the water quality and hydrographic data and ultimately output as a data table.
}
