% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAhydro}
\alias{.loadNCCAhydro}
\title{Load and join hydrographic and Secchi data for NCCA 2010 and 2015}
\usage{
.loadNCCAhydro(
  NCCAhydrofiles2010,
  NCCAsites2010,
  NCCAhydrofile2015,
  NCCAsites2015,
  NCCAsecchifile2015,
  namingFile,
  n_max = Inf
)
}
\arguments{
\item{NCCAhydrofiles2010}{a string specifying the URL for the 2010 hydrographic data}

\item{NCCAsites2010}{a string specifying the URL for the 2010 site data}

\item{NCCAhydrofile2015}{a string specifying the URL for the 2015 hydrographic data}

\item{NCCAsites2015}{a string specifying the URL for the 2015 site data}

\item{NCCAsecchifile2015}{a string specifying the URL for the 2015 Secchi data}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{n_max}{Number of rows to read in from the data file (this is just for testing purposes)}
}
\value{
dataframe
}
\description{
\code{.loadNCCAhydro} returns a dataframe of all of the hydrographic data relating to NCCA 2010 and 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
