% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAhydro2010}
\alias{.loadNCCAhydro2010}
\title{Load and join NCCA 2010 hydrographic data from csv files}
\usage{
.loadNCCAhydro2010(
  NCCAhydrofiles2010,
  NCCAsites2010,
  namingFile,
  n_max = n_max
)
}
\arguments{
\item{filepath}{a string specifying the filepath of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCAhydro2010} returns a dataframe of all of the hydrographic data relating to NCCA 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
