% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAhydro2015}
\alias{.loadNCCAhydro2015}
\title{Load and join NCCA 2015 hydrographic data from online csv files}
\usage{
.loadNCCAhydro2015(NCCAhydrofile2015, NCCAsites2015, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAhydrofile2015}{a string specifying the URL for the data}

\item{NCCAsites2015}{a string specifying the URL for the site data}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{n_max}{Number of rows to read in from the data file (this is just for testing purposes)}
}
\value{
dataframe
}
\description{
\code{.loadNCCAhydro2015} returns a dataframe of all of the hydrographic data relating to NCCA 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
