% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAhydrological.R
\name{.loadNCCAsecchi2015}
\alias{.loadNCCAsecchi2015}
\title{Load and join secchi data for NCCA 2015 from csv files}
\usage{
.loadNCCAsecchi2015(NCCAsecchifile2015, NCCAsites2015, namingFile, n_max = Inf)
}
\arguments{
\item{filepath}{a string specifying the filepath of the data}
}
\value{
dataframe of the fully joined secchi data from NCCA 2015
}
\description{
\code{.loadNCCAsecchi2015} returns a dataframe of all of the joined secchi data relating to NCCA 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
