% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.loadNCCAwq2010}
\alias{.loadNCCAwq2010}
\title{Read in NCCA water chemistry from 2010}
\usage{
.loadNCCAwq2010(NCCAwq2010, NCCAsites2010, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAwq2010}{a string specifying the URL for the data}

\item{NCCAsites2010}{a string specifying the URL for the site data}

\item{namingFile}{a string specifying the URL for the analyte naming file}

\item{n_max}{Number of rows to read in from the data file (this is just for testing purposes)}
}
\value{
dataframe
}
\description{
\code{.loadNCCAwq2010} returns water chemistry data measured during the NCCA study in 2010
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
