% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAwaterChemistry.R
\name{.loadNCCAwq2015}
\alias{.loadNCCAwq2015}
\title{Read in all NCCA water chemistry from 2015}
\usage{
.loadNCCAwq2015(NCCAwq2015, NCCAsites2015, namingFile, n_max = Inf)
}
\arguments{
\item{NCCAwq2015}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.loadNCCAwq2015} returns water chemistry data measured during the NCCA study in 2015
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
