% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCCAsites.R
\name{.readNCCASites}
\alias{.readNCCASites}
\title{Read in all NCCA site data}
\usage{
.readNCCASites(ncca2010sites, ncca2015sites)
}
\arguments{
\item{filepath}{a string specifying the directory of the data}
}
\value{
dataframe
}
\description{
\code{.readNCCASites} returns spatial data relating to study sites
}
\details{
This is a hidden function, this should be used for development purposes only, users will only call
this function implicitly when assembling their full water quality dataset
}
